feat: small forms in new format (#20) (#28)
* feat: small forms in new format (#20)

* new basic assessment and other changes

* Apply "async" review suggestions


* reverting review suggestions
