+++ action-server/covidflow/actions/home_assistance_form.py
@@ -1,20 +1,53 @@
-from typing import Any, Dict, List, Optional, Text, Union
+from typing import Any, Dict, List, Text
 
-from rasa_sdk import Tracker
-from rasa_sdk.events import EventType
+import structlog
+from rasa_sdk import Action, Tracker
+from rasa_sdk.events import EventType, SlotSet
 from rasa_sdk.executor import CollectingDispatcher
-from rasa_sdk.forms import FormAction
+from rasa_sdk.forms import REQUESTED_SLOT
 
-from covidflow.constants import HAS_ASSISTANCE_SLOT, PROVINCE_SLOT, PROVINCES_WITH_211
-
-from .lib.form_helper import (
-    request_next_slot,
-    validate_boolean_slot,
-    yes_no_nlu_mapping,
+from covidflow.constants import (
+    HAS_ASSISTANCE_SLOT,
+    PROVINCE_SLOT,
+    PROVINCES_WITH_211,
+    SKIP_SLOT_PLACEHOLDER,
 )
+
 from .lib.log_util import bind_logger
 
 FORM_NAME = "home_assistance_form"
+VALIDATE_ACTION_NAME = f"validate_{FORM_NAME}"
+
+logger = structlog.get_logger()
+
+
+class ValidateHomeAssistanceForm(Action):
+    def name(self) -> Text:
+        return VALIDATE_ACTION_NAME
+
+    def run(
+        self, dispatcher: CollectingDispatcher, tracker: Tracker, domain: Dict
+    ) -> List[EventType]:
+        bind_logger(tracker)
+
+        extracted_slots: Dict[Text, Any] = tracker.form_slots_to_validate()
+
+        if extracted_slots == {}:  # On activation
+            if not _has_211(tracker):
+                _display_recommendations(False, dispatcher, tracker)
+
+                return [
+                    SlotSet(REQUESTED_SLOT, None),
+                    SlotSet(HAS_ASSISTANCE_SLOT, SKIP_SLOT_PLACEHOLDER),
+                ]
+            return []
+
+        if extracted_slots[HAS_ASSISTANCE_SLOT] is True:
+            _display_recommendations(False, dispatcher, tracker)
+        else:
+            _display_recommendations(True, dispatcher, tracker)
+
+        return [SlotSet(HAS_ASSISTANCE_SLOT, extracted_slots[HAS_ASSISTANCE_SLOT])]
 
 
 def _has_211(tracker: Tracker) -> bool:
@@ -22,64 +55,20 @@ def _has_211(tracker: Tracker) -> bool:
     return province in PROVINCES_WITH_211
 
 
-class HomeAssistanceForm(FormAction):
-    def name(self) -> Text:
-
-        return FORM_NAME
-
-    async def run(
-        self, dispatcher, tracker, domain,
-    ):
-        bind_logger(tracker)
-        return await super().run(dispatcher, tracker, domain)
-
-    @staticmethod
-    def required_slots(tracker: Tracker) -> List[Text]:
-        if _has_211(tracker):
-            return [HAS_ASSISTANCE_SLOT]
-
-        return []
-
-    def slot_mappings(self) -> Dict[Text, Union[Dict, List[Dict]]]:
-        return {
-            HAS_ASSISTANCE_SLOT: yes_no_nlu_mapping(self),
-        }
-
-    def request_next_slot(
-        self,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict[Text, Any],
-    ) -> Optional[List[EventType]]:
-        return request_next_slot(self, dispatcher, tracker, domain, None)
-
-    @validate_boolean_slot
-    def validate_has_assistance(
-        self,
-        value: Union[bool, Text],
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict[Text, Any],
-    ) -> Dict[Text, Any]:
-        return {}
-
-    def submit(
-        self,
-        dispatcher: CollectingDispatcher,
-        tracker: Tracker,
-        domain: Dict[Text, Any],
-    ) -> List[Dict]:
-
-        if _has_211(tracker) and not tracker.get_slot(HAS_ASSISTANCE_SLOT):
-            dispatcher.utter_message(template="utter_check_delivery_services")
-            if tracker.get_slot(PROVINCE_SLOT) == "qc":
-                dispatcher.utter_message(template="utter_may_call_211_qc")
-            else:
-                dispatcher.utter_message(template="utter_may_call_211_other_provinces")
-            dispatcher.utter_message(template="utter_explain_211")
+def _display_recommendations(
+    propose_211: bool, dispatcher: CollectingDispatcher, tracker: Tracker
+) -> None:
+    if propose_211:
+        dispatcher.utter_message(template="utter_home_assistance_offer_211_true_1")
+        if tracker.get_slot(PROVINCE_SLOT) == "qc":
+            dispatcher.utter_message(
+                template="utter_home_assistance_offer_211_true_2_qc"
+            )
         else:
-            dispatcher.utter_message(template="utter_remind_delivery_services")
-
-        dispatcher.utter_message(template="utter_remind_pharmacist_services")
-
-        return []
+            dispatcher.utter_message(
+                template="utter_home_assistance_offer_211_true_2_other"
+            )
+        dispatcher.utter_message(template="utter_home_assistance_offer_211_true_3")
+    else:
+        dispatcher.utter_message(template="utter_home_assistance_offer_211_false")
+    dispatcher.utter_message(template="utter_home_assistance_final")
